Fixed up for new codegen

+++ SpringProject/manipulability.m
@@ -10,7 +10,7 @@ state(19:36) = zeros(1,18);
 [~,SIG_BR,~] = svd(GeoJc_BR);
 [~,SIG_BL,~] = svd(GeoJc_BL);
 
-sig_size = size(SIG_FR, 1);
+sig_size = uint16(size(SIG_FR, 1));
 
 sigFR = zeros(6,1);
 sigFL = zeros(6,1);
